# shared/openapi/openapi.json
{
  "components": {
    "schemas": {
      "AcceptedEvent": {
        "properties": {
          "device_id": {
            "title": "Device Id",
            "type": "integer"
          },
          "seq": {
            "title": "Seq",
            "type": "integer"
          },
          "server_seq": {
            "title": "Server Seq",
            "type": "integer"
          },
          "was_duplicate": {
            "title": "Was Duplicate",
            "type": "boolean"
          }
        },
        "required": [
          "device_id",
          "seq",
          "server_seq",
          "was_duplicate"
        ],
        "title": "AcceptedEvent",
        "type": "object"
      },
      "AssignmentCreate": {
        "additionalProperties": false,
        "properties": {
          "cell_ids": {
            "items": {
              "type": "integer"
            },
            "minItems": 1,
            "title": "Cell Ids",
            "type": "array"
          },
          "incident_id": {
            "minimum": 0.0,
            "title": "Incident Id",
            "type": "integer"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "team_id": {
            "maximum": 255.0,
            "minimum": 0.0,
            "title": "Team Id",
            "type": "integer"
          }
        },
        "required": [
          "team_id",
          "cell_ids",
          "incident_id"
        ],
        "title": "AssignmentCreate",
        "type": "object"
      },
      "AssignmentCreated": {
        "properties": {
          "already_held": {
            "items": {
              "$ref": "#/components/schemas/HeldCell"
            },
            "title": "Already Held",
            "type": "array"
          },
          "assignment_id": {
            "title": "Assignment Id",
            "type": "integer"
          },
          "server_seq": {
            "title": "Server Seq",
            "type": "integer"
          }
        },
        "required": [
          "assignment_id",
          "server_seq",
          "already_held"
        ],
        "title": "AssignmentCreated",
        "type": "object"
      },
      "AssignmentRevoked": {
        "properties": {
          "server_seq": {
            "title": "Server Seq",
            "type": "integer"
          }
        },
        "required": [
          "server_seq"
        ],
        "title": "AssignmentRevoked",
        "type": "object"
      },
...
      "DeltaPage": {
        "properties": {
          "current_server_seq": {
            "title": "Current Server Seq",
            "type": "integer"
          },
          "events": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Events",
            "type": "array"
          },
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "next_since": {
            "title": "Next Since",
            "type": "integer"
          }
        },
        "required": [
          "events",
          "next_since",
          "has_more",
          "current_server_seq"
        ],
        "title": "DeltaPage",
        "type": "object"
      },
      "DisasterType": {
        "description": "docs/08 \u00a72.1 event 1; the incident types named in PS-B1.",
        "enum": [
          "LANDSLIDE",
          "BUILDING_COLLAPSE",
          "OTHER"
        ],
        "title": "DisasterType",
        "type": "string"
      },
...
      "EventBatch": {
        "additionalProperties": false,
        "properties": {
          "batch_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Id"
          },
          "events": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "minItems": 1,
            "title": "Events",
            "type": "array"
          },
          "mac": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mac"
          },
          "origin_device_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Device Id"
          }
        },
        "required": [
          "events"
        ],
        "title": "EventBatch",
        "type": "object"
      },
      "GridCreate": {
        "additionalProperties": false,
        "properties": {
          "bearing_deg": {
            "default": 0.0,
            "exclusiveMaximum": 360.0,
            "minimum": 0.0,
            "title": "Bearing Deg",
            "type": "number"
          },
          "cell_size_m": {
            "exclusiveMinimum": 0.0,
            "title": "Cell Size M",
            "type": "number"
          },
          "cols": {
            "maximum": 8191.0,
            "minimum": 1.0,
            "title": "Cols",
            "type": "integer"
          },
          "origin_lat": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Origin Lat",
            "type": "number"
          },
          "origin_lon": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Origin Lon",
            "type": "number"
          },
          "rows": {
            "maximum": 8191.0,
            "minimum": 1.0,
            "title": "Rows",
            "type": "integer"
          }
        },
        "required": [
          "origin_lat",
          "origin_lon",
          "cell_size_m",
          "rows",
          "cols"
        ],
        "title": "GridCreate",
        "type": "object"
      },
      "GridCreated": {
        "properties": {
          "cell_count": {
            "title": "Cell Count",
            "type": "integer"
          },
          "incident_id": {
            "title": "Incident Id",
            "type": "integer"
          },
          "server_seq": {
            "title": "Server Seq",
            "type": "integer"
          }
        },
        "required": [
          "incident_id",
          "cell_count",
          "server_seq"
        ],
        "title": "GridCreated",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HeldCell": {
        "properties": {
          "cell_index": {
            "title": "Cell Index",
            "type": "integer"
          },
          "team_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Id"
          }
        },
        "required": [
          "cell_index",
          "team_id"
        ],
        "title": "HeldCell",
        "type": "object"
      },
      "IncidentCreate": {
        "additionalProperties": false,
        "properties": {
          "disaster_type": {
            "$ref": "#/components/schemas/DisasterType"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "origin_lat": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Origin Lat",
            "type": "number"
          },
          "origin_lon": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Origin Lon",
            "type": "number"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          }
        },
        "required": [
          "name",
          "disaster_type",
          "origin_lat",
          "origin_lon",
          "started_at"
        ],
        "title": "IncidentCreate",
        "type": "object"
      },
      "IncidentCreated": {
        "properties": {
          "incident_id": {
            "title": "Incident Id",
            "type": "integer"
          },
          "server_seq": {
            "title": "Server Seq",
            "type": "integer"
          }
        },
        "required": [
          "incident_id",
          "server_seq"
        ],
        "title": "IncidentCreated",
        "type": "object"
      },
      "IngestResult": {
        "properties": {
          "accepted": {
            "items": {
              "$ref": "#/components/schemas/AcceptedEvent"
            },
            "title": "Accepted",
            "type": "array"
          },
          "rejected": {
            "items": {
              "$ref": "#/components/schemas/RejectedEvent"
            },
            "title": "Rejected",
            "type": "array"
          },
          "server_seq_high": {
            "title": "Server Seq High",
            "type": "integer"
          }
        },
        "required": [
          "accepted",
          "rejected",
          "server_seq_high"
        ],
        "title": "IngestResult",
        "type": "object"
      },
...
      "RejectedEvent": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "device_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "seq": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seq"
          }
        },
        "required": [
          "code",
          "retryable",
          "message"
        ],
        "title": "RejectedEvent",
        "type": "object"
      },
...
      "TeamCreate": {
        "additionalProperties": false,
        "properties": {
          "agency_id": {
            "maximum": 15.0,
            "minimum": 0.0,
            "title": "Agency Id",
            "type": "integer"
          },
          "agency_type": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Agency Type",
            "type": "string"
          },
          "callsign": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Callsign",
            "type": "string"
          },
          "incident_id": {
            "minimum": 0.0,
            "title": "Incident Id",
            "type": "integer"
          },
          "team_id": {
            "maximum": 255.0,
            "minimum": 0.0,
            "title": "Team Id",
            "type": "integer"
          }
        },
        "required": [
          "team_id",
          "callsign",
          "agency_id",
          "agency_type",
          "incident_id"
        ],
        "title": "TeamCreate",
        "type": "object"
      },
      "TeamCreated": {
        "properties": {
          "server_seq": {
            "title": "Server Seq",
            "type": "integer"
          },
          "team_id": {
            "title": "Team Id",
            "type": "integer"
          }
        },
        "required": [
          "team_id",
          "server_seq"
        ],
        "title": "TeamCreated",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Multi-agency search-and-rescue coordination. REST is authoritative for all durable writes and reads.",
    "title": "RescueNet AI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/assignments": {
      "post": {
        "operationId": "issue_assignment_v1_assignments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignmentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Issue an assignment",
        "tags": [
          "assignment"
        ]
      }
    },
    "/v1/assignments/{assignment_id}": {
      "delete": {
        "operationId": "revoke_assignment_v1_assignments__assignment_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "assignment_id",
            "required": true,
            "schema": {
              "title": "Assignment Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "reason",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reason"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentRevoked"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke an assignment",
        "tags": [
          "assignment"
        ]
      }
    },
...
    "/v1/events": {
      "get": {
        "operationId": "read_events_v1_events_get",
        "parameters": [
          {
            "description": "Cursor \u2014 exclusive lower bound on server_seq",
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Cursor \u2014 exclusive lower bound on server_seq",
              "minimum": 0,
              "title": "Since",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "entity_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Type"
            }
          },
          {
            "in": "query",
            "name": "entity_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Id"
            }
          },
          {
            "in": "query",
            "name": "team_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Team Id"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeltaPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delta sync and audit history",
        "tags": [
          "events"
        ]
      },
      "post": {
        "description": "Partial acceptance: 200 even when some events are rejected (AP-04).",
        "operationId": "upload_events_v1_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Batch event upload",
        "tags": [
          "events"
        ]
      }
    },
    "/v1/incidents": {
      "post": {
        "operationId": "declare_incident_v1_incidents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Declare an incident",
        "tags": [
          "incident"
        ]
      }
    },
    "/v1/incidents/{incident_id}/export": {
      "get": {
        "operationId": "export_v1_incidents__incident_id__export_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "geojson",
              "pattern": "^(geojson|csv)$",
              "title": "Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "GeoJSON / CSV export",
        "tags": [
          "incident"
        ]
      }
    },
    "/v1/incidents/{incident_id}/grid": {
      "post": {
        "operationId": "generate_grid_v1_incidents__incident_id__grid_post",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GridCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GridCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate the grid",
        "tags": [
          "incident"
        ]
      }
    },
    "/v1/incidents/{incident_id}/snapshot": {
      "get": {
        "operationId": "snapshot_v1_incidents__incident_id__snapshot_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "command",
              "pattern": "^(field|command)$",
              "title": "Scope",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Snapshot V1 Incidents  Incident Id  Snapshot Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Initial load",
        "tags": [
          "incident"
        ]
      }
    },
    "/v1/teams": {
      "post": {
        "operationId": "register_team_v1_teams_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register a team",
        "tags": [
          "team"
        ]
      }
    },
...
  }
}